+++ tests/serialization/test_result.py
@@ -2,9 +2,9 @@ import marshmallow
 import pytest
 
 import prefect
-from prefect.engine.result import Result, NoResult
+from prefect.engine.result import Result, NoResult, NoResultType
 from prefect.engine.result_handlers import JSONResultHandler
-from prefect.serialization.result import ResultSchema, NoResultSchema
+from prefect.serialization.result import ResultSchema, NoResultSchema, StateResultSchema
 
 
 def test_basic_result_serializes():
@@ -17,8 +17,8 @@ def test_basic_result_serializes():
 
 
 def test_basic_noresult_serializes():
-    r = NoResult()
-    handled = ResultSchema().dump(r)
+    r = NoResult
+    handled = NoResultSchema().dump(r)
     version = handled.pop("__version__")
     assert version == prefect.__version__
     assert handled == {}
@@ -32,9 +32,16 @@ def test_basic_result_deserializes():
     assert r.result_handler is None
 
 
+def test_roundtrip_for_results():
+    r = Result(value={"y": 4})
+    s = ResultSchema().load(ResultSchema().dump(r))
+    assert r == s
+
+
 def test_basic_noresult_deserializes():
     r = NoResultSchema().load({})
-    assert isinstance(r, NoResult)
+    assert r == NoResult
+    assert isinstance(r, NoResultType)
 
 
 def test_result_serializes_result_handlers():
@@ -43,7 +50,21 @@ def test_result_serializes_result_handlers():
     assert handled["result_handler"]["type"] == "JSONResultHandler"
 
 
-def test_result_allows_none_value():
-    r = Result(value=None)
-    handled = ResultSchema().dump(r)
-    assert handled["value"] is None
+def test_result_allows_none_value_and_handler():
+    schema = ResultSchema()
+    r = Result(value=None, result_handler=None)
+    handled = schema.load(schema.dump(r))
+    assert handled.value is None
+    assert handled.result_handler is None
+
+
+@pytest.mark.parametrize("obj", [Result(value=19), NoResult])
+def test_state_result_schema_chooses_schema(obj):
+    schema = StateResultSchema()
+    assert type(schema.load(schema.dump(obj))) == type(obj)
+
+
+def test_value_raises_error_on_dump_if_not_valid_json():
+    r = Result(value={"x": {"y": {"z": lambda: 1}}})
+    with pytest.raises(marshmallow.exceptions.ValidationError):
+        StateResultSchema().dump(r)